requirements.txt
requests>=2.28.0
pystray>=0.19.0
Pillow>=9.0.0
matplotlib>=3.7.0
cairosvg>=2.5.0
pyinstaller>=6.0.0
